EduVisual/EduVisual/requirement.txt
python-dotenv
requests
gtts
pygame